Properly inject Capacitor js code inside head (#1106)
